Move entry-wide contexts and state back to Entry
Pass state as props to `Chapter`
